AG-11136 - Rename regions file
