+++ packages/ag-charts-community/src/chart/interaction/regions.ts
